:truck: api(devcomp): move unit test to devcomp context
