+++ api/tests/devcomp/unit/application/campaign-participations/index_test.js
@@ -1,6 +1,6 @@
-import { campaignParticipationController } from '../../../../src/devcomp/application/campaign-participations/campaign-participation-controller.js';
-import * as moduleUnderTest from '../../../../src/devcomp/application/campaign-participations/campaign-participation-route.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import { campaignParticipationController } from '../../../../../src/devcomp/application/campaign-participations/campaign-participation-controller.js';
+import * as moduleUnderTest from '../../../../../src/devcomp/application/campaign-participations/campaign-participation-route.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-participation-router ', function () {
   describe('GET /api/campaign-participations/{id}/trainings', function () {
