+++ api/tests/devcomp/unit/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer_test.js
@@ -1,6 +1,6 @@
-import { TutorialEvaluation } from '../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
-import * as serializer from '../../../../../src/devcomp/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { TutorialEvaluation } from '../../../../../../src/devcomp/domain/models/TutorialEvaluation.js';
+import * as serializer from '../../../../../../src/devcomp/infrastructure/serializers/jsonapi/tutorial-evaluation-serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | tutorial-evaluation-serializer', function () {
   describe('#serialize', function () {
